Remove reference to domain from main screen view setting.
