+++ src/main/java/oripa/domain/paint/copypaste/SelectionOriginHolder.java
@@ -1,32 +1,37 @@
+/**
+ * ORIPA - Origami Pattern Editor 
+ * Copyright (C) 2013-     ORIPA OSS Project  https://github.com/oripa/oripa
+ * Copyright (C) 2005-2009 Jun Mitani         http://mitani.cs.tsukuba.ac.jp/
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 package oripa.domain.paint.copypaste;
 
 import javax.vecmath.Vector2d;
 
 import oripa.domain.paint.PaintContext;
 
-public class SelectionOriginHolder {
+/**
+ * @author OUCHI Koji
+ *
+ */
+public interface SelectionOriginHolder {
 
-	public SelectionOriginHolder() {
-	}
+	void setOrigin(Vector2d p);
 
-	private Vector2d origin = null;
+	void resetOrigin(PaintContext context);
 
-	public void setOrigin(final Vector2d p) {
-		origin = p;
-	}
-
-	public void resetOrigin(final PaintContext context) {
-		if (origin == null) {
-			if (context.getLineCount() > 0) {
-				origin = context.getLine(0).p0;
-			}
-		}
-	}
-
-	public Vector2d getOrigin(final PaintContext context) {
-		resetOrigin(context);
-
-		return origin;
-	}
+	Vector2d getOrigin(PaintContext context);
 
 }
